css/css-will-change/will-change-002.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>will-change invalid keywords case-insensitivity</title>
<link rel="help" href="https://drafts.csswg.org/css-will-change-1/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
/*
 * When parsing the `will-change` property, a `<custom-ident>` value must be considered invalid if it is an ASCII case-insensitive match for the keywords `will-change`, `none`, `all`, `auto`, `scroll-position`, or `contents`.
 */
test(() => {
  assert_false(CSS.supports("will-change", "WILL-CHANGE"), "WILL-CHANGE should not be supported");
  assert_false(CSS.supports("will-change", "NONE"), "NONE should not be supported");
  assert_false(CSS.supports("will-change", "ALL"), "ALL should not be supported");
  assert_false(CSS.supports("will-change", "AUTO, transform"), "AUTO, transform should not be supported");
}, "will-change invalid keywords case-insensitivity");
</script>
